Added copywin method

+++ src/main/java/io/webfolder/curses4j/CursesWindow.java
@@ -97,7 +97,7 @@ class CursesWindow {
             int src_tc, int dst_tr, int dst_tc, int dst_br,
             int dst_bc, int _overlay);
 
-    native static int curses4j_dupwin(long win);
+    native static long curses4j_dupwin(long win);
 
     native int curses4j_winsertln(long peer);
